Replaced finnhub-stream.py with FinnhubProducer.py
